+++ R/dvsfromev.R
@@ -10,13 +10,16 @@
 #'   explanatory variable may be continuous or categorical.
 #' @param transformtype Set of transformation types to be used.
 #' @param allsplines Logical. Keep all spline transformations.
-#' @param dir Directory to which Maxent runs of spline transformations are
-#'   written
+#' @param algorithm Character string matching either "maxent" or "LR".
+#' @param quiet Logical. Suppress progress messages from spline pre-seletion?
 #'
 #' @return Dataframe with one column for each DV.
+#'
+#' @keywords internal
+#' @noRd
 
 
-.dvsfromev <- function(df, transformtype, allsplines, dir) {
+.dvsfromev <- function(df, transformtype, allsplines, algorithm, quiet) {
 
   rv <- df[, 1]
   ev <- df[, 2]
@@ -24,10 +27,6 @@
   storage <- list()
 
   if (class(ev) %in% c("numeric", "integer")) {
-
-    if (any(c("HF", "HR", "T") %in% transformtype) && allsplines == F) {
-      evdir <- .dirpath.create(dir, evname)
-    }
 
     if ("L" %in% transformtype) {
       tfunction <- .transfL(ev)
@@ -42,7 +41,7 @@
     if ("D" %in% transformtype) {
       for (i in c(0.5, 1, 2)) {
         tfunction <- .transfD(ev, rv, i)
-        storage[[paste0(evname, "_D", i, "_transf")]] <- tfunction
+        storage[[paste0(evname, "_D", sub("[.]", "", i), "_transf")]] <- tfunction
       }
     }
 
@@ -54,14 +53,15 @@
         tfunction <- .transfSpline(ev, k, type = "HF")
         splall[[paste0(evname, "_HF", i, "_transf")]] <- tfunction
       }
-      if (allsplines == T) {
+      if (allsplines == TRUE) {
         storage <- c(storage, splall)
       } else {
-        hfdir <- .dirpath.create(evdir, "HF")
-        message(paste0("Pre-selecting forward hinge transformations of ", evname))
+        if (quiet == FALSE) {
+          message(paste0("Pre-selecting forward hinge transformations of ", evname))
+          }
         dvs <- lapply(splall, function(x) {x(ev)})
         names(dvs) <- gsub("_transf", "", names(splall))
-        selected <- .splselect(rv, dvs, hfdir)
+        selected <- .splselect(rv, dvs, algorithm)
         if (length(selected) > 0) {
           storage <- c(storage, splall[paste0(selected, "_transf")])
         }
@@ -76,14 +76,15 @@
         tfunction <- .transfSpline(ev, k, type = "HR")
         splall[[paste0(evname, "_HR", i, "_transf")]] <- tfunction
       }
-      if (allsplines == T) {
+      if (allsplines == TRUE) {
         storage <- c(storage, splall)
       } else {
-        hrdir <- .dirpath.create(evdir, "HR")
-        message(paste0("Pre-selecting reverse hinge transformations of ", evname))
+        if (quiet == FALSE) {
+          message(paste0("Pre-selecting reverse hinge transformations of ", evname))
+        }
         dvs <- lapply(splall, function(x) {x(ev)})
         names(dvs) <- gsub("_transf", "", names(splall))
-        selected <- .splselect(rv, dvs, hrdir)
+        selected <- .splselect(rv, dvs, algorithm)
         if (length(selected) > 0) {
           storage <- c(storage, splall[paste0(selected, "_transf")])
         }
@@ -98,14 +99,15 @@
         tfunction <- .transfSpline(ev, k, type = "T")
         splall[[paste0(evname, "_T", i, "_transf")]] <- tfunction
       }
-      if (allsplines == T) {
+      if (allsplines == TRUE) {
         storage <- c(storage, splall)
       } else {
-        tdir <- .dirpath.create(evdir, "T")
-        message(paste0("Pre-selecting threshold transformations of ", evname))
+        if (quiet == FALSE) {
+          message(paste0("Pre-selecting threshold transformations of ", evname))
+        }
         dvs <- lapply(splall, function(x) {x(ev)})
         names(dvs) <- gsub("_transf", "", names(splall))
-        selected <- .splselect(rv, dvs, tdir)
+        selected <- .splselect(rv, dvs, algorithm)
         if (length(selected) > 0) {
           storage <- c(storage, splall[paste0(selected, "_transf")])
         }
@@ -119,7 +121,8 @@
     if ("B" %in% transformtype) {
       for (i in levels(ev)) {
         tfunction <- .transfB(ev, i)
-        storage[[paste0(evname, "_B", i, "_transf")]] <- tfunction
+        levelname <- make.names(i, allow_ = FALSE)
+        storage[[paste0(evname, "_B", levelname, "_transf")]] <- tfunction
       }
     }
   }